backend/backend/pyproject.toml
[project]
name = "ostaffsync"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "duckdb>=1.4.1",
    "numpy>=2.3.3",
    "pandas>=2.3.3",
    "reflex-hosting-cli>=0.1.56",
    "reflex-motion>=0.0.1",
]
